let unit prices accept floating value

+++ src-tauri/src/command/journal.rs
@@ -20,7 +20,7 @@ pub struct JournalRecordData {
     supplier_id: String,
     supplier_name: String,
     unit_name: String,
-    unit_price: u32,
+    unit_price: f64,
     quantity: f64,
 }
